MAT-7422 refactored code
