+++ src/CqlBuilderPanel/definitionsSection/DefinitionSection.tsx
@@ -10,7 +10,7 @@ import {
 import "./Definitions.scss";
 import { DefinitionSectionSchemaValidator } from "../../validations/DefinitionSectionSchemaValidator";
 import ExpressionEditor from "./expressionSection/ExpressionEditor";
-import { CqlBuilderAvailableLookupData } from "./expressionSection/ExpressionEditorHelper";
+import { CqlBuilderAvailableLookupData } from "../../model/CqlBuilderLookup";
 
 export interface Definition {
   definitionName?: string;
